Fix to support the value/valueSpecified properties used by XML in Json.
